Graphics Demo's War Against Eyes is Officially Over!
Dark mode is here, and oh boy did I have to do some questionable css to do it (blood to the css variables god).

Half the bulma css might be semi-broken, and this might not work in older browsers now, but hey, isn't that the price of progress? At least the color.js is more streamlined.